ifdown: Introduce _UnconfigureNIC() helper method
This method takes a NIC object, creates the proper invironment
and invokes the ifdown script.

It takes an extra boolean argument 'only_local' which is passed
to ifdown script as the second positional argument. It states
if the external script should wipe out every configuration
related to the NIC or should only clean up local thinks, e.g.
during a successful migration, we are going to invoke ifdown
on the source node but we do not want to purge any DDNS
configuration related to the instance's NICs.

Signed-off-by: Dimitris Aragiorgis <dimitris.aragiorgis@gmail.com>
Signed-off-by: Yiannis Tsiouris <tsiou@grnet.gr>