cfgupgrade: Switch to non base64 encoded networks
Use 01 bitarrays instead of base64 encoded strings for storing
network reservations.

cfgupgrade now runs through gnt-cluster upgrade and passing
options explicitly it not possible.

Still we add a --tob64 option if one want to revert to base64
encoded networks. Note that if the upgrade is done already
we have to pass --force to run it again (it should be a no-op
for all other actions).

Signed-off-by: Dimitris Aragiorgis <dimitris.aragiorgis@gmail.com>