Add clear OS params options in gnt-instance modify
This patch adds two new options in 'gnt-instance modify' which allow the
user to clear all current public/private OS parameters of an instance.
This might be useful for OS providers that consider no parameters as
valid or, more commonly, for changing OS provider (and parameters)
before performing a reinstall. E.g.:

$ gnt-instance modify --clear-os-parameters <instance_name>

or

$ gnt-instance modify --clear-os-parameters-private <instance_name>

Signed-off-by: Yiannis Tsiouris <tsiou@grnet.gr>
Reviewed-by: Brian Foley <bpfoley@google.com>